chore(*): Reorganized all views structure
