+++ src/views/data-workflows/gcs-to-gcs/GcsToGcsRuns.vue
@@ -1,17 +1,17 @@
 <template>
 	<div>
 		<DataManagementHeader :tabsItems="tabsItems" :headerTitle="headerTitle"></DataManagementHeader>
-		<GcsToGcsConfs />
+		<GcsToGcsRuns />
 	</div>
 </template>
 
 <script>
-import GcsToGcsConfs from '../components/data-workflows/gcs-to-gcs/GcsToGcsConfs';
-import DataManagementHeader from '../components/app/headers/DataManagementHeader';
+import GcsToGcsRuns from '../../../components/data-workflows/gcs-to-gcs/GcsToGcsRuns';
+import DataManagementHeader from '../../../components/app/headers/DataManagementHeader';
 
 export default {
 	components: {
-		GcsToGcsConfs,
+		GcsToGcsRuns,
 		DataManagementHeader
 	},
 	data() {
