+++ src/views/data-workflows/table-to-storage/TableToStorageRuns.vue
@@ -6,8 +6,8 @@
 </template>
 
 <script>
-import RunsListing from '../components/data-workflows/table-to-storage/RunsListing';
-import DataManagementHeader from '../components/app/headers/DataManagementHeader';
+import RunsListing from '../../../components/data-workflows/table-to-storage/RunsListing';
+import DataManagementHeader from '../../../components/app/headers/DataManagementHeader';
 
 export default {
 	components: {
