+++ src/views/data-workflows/tables-to-tables/TablesToTablesRun.vue
@@ -1,17 +1,17 @@
 <template>
 	<div>
 		<DataManagementHeader :tabsItems="tabsItems" :headerTitle="headerTitle"></DataManagementHeader>
-		<TablesToTablesConf />
+		<TablesToTablesRun />
 	</div>
 </template>
 
 <script>
-import TablesToTablesConf from '../components/data-workflows/tables-to-tables/TablesToTablesConf';
-import DataManagementHeader from '../components/app/headers/DataManagementHeader';
+import TablesToTablesRun from '../../../components/data-workflows/tables-to-tables/TablesToTablesRun';
+import DataManagementHeader from '../../../components/app/headers/DataManagementHeader';
 
 export default {
 	components: {
-		TablesToTablesConf,
+		TablesToTablesRun,
 		DataManagementHeader
 	},
 	data() {
