+++ src/views/data-workflows/vm-launcher/VmLauncherRuns.vue
@@ -6,8 +6,8 @@
 </template>
 
 <script>
-import VmLauncherRuns from '../components/data-workflows/vm-launcher/VmLauncherRuns';
-import DataManagementHeader from '../components/app/headers/DataManagementHeader';
+import VmLauncherRuns from '../../../components/data-workflows/vm-launcher/VmLauncherRuns';
+import DataManagementHeader from '../../../components/app/headers/DataManagementHeader';
 
 export default {
 	components: {
